+++ src/components/Buttons/TextButton.test.tsx
@@ -1,10 +1,10 @@
-import Button from '@components/Buttons/Button';
+import TextButton from '@components/Buttons/TextButton';
 import { render } from '@tests/test-utils';
 
 describe('Button Component', () => {
-  it('Basic Button test', () => {
+  it('Text Button test', () => {
     const onClick = jest.fn();
-    const { getByRole } = render(<Button onClick={onClick}>Button</Button>);
+    const { getByRole } = render(<TextButton text='button' />);
     const button = getByRole('button');
     button.click();
     expect(onClick).toBeCalled();
